Revert "feat(core/github-api): Centralised Github Fetching And Rate Limiting using API v3 (#2181)"
This reverts commit ac084e80ecbde461b91931c7c2f62db877686821.
