+++ src/core/contrib.js
@@ -4,8 +4,8 @@
 // #gh-commenters: people having contributed comments to issues.
 // #gh-contributors: people whose PR have been merged.
 // Spec editors get filtered out automatically.
-import { fetchIndex, getRateLimit, githubRequestHeaders } from "./github-api";
 import { flatten, joinAnd } from "./utils";
+import { fetchIndex } from "./github";
 import { pub } from "./pubsubhub";
 export const name = "core/contrib";
 
@@ -24,11 +24,9 @@ function findUserURLs(...thingsWithUsers) {
 }
 
 async function toHTML(urls, editors, element, headers) {
-  const args = await Promise.all(
-    urls.map(url => fetch(new Request(url, { headers })).then(r => r.json()))
-  );
+  const args = await Promise.all(urls.map(url => fetch(url, { headers })));
   const names = args
-    .map(user => user.name || user.login)
+    .map(([user]) => user.name || user.login)
     .filter(name => !editors.includes(name))
     .sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
   element.textContent = joinAnd(names);
@@ -41,7 +39,13 @@ export async function run(conf) {
   if (!ghCommenters && !ghContributors) {
     return;
   }
-  const { githubAPI } = conf;
+  const headers = {};
+  const { githubAPI, githubUser, githubToken } = conf;
+  if (githubUser && githubToken) {
+    const credentials = btoa(`${githubUser}:${githubToken}`);
+    const Authorization = `Basic ${credentials}`;
+    Object.assign(headers, { Authorization });
+  }
   if (!githubAPI) {
     const msg =
       "Requested list of contributors and/or commenters from GitHub, but " +
@@ -49,8 +53,7 @@ export async function run(conf) {
     pub("error", msg);
     return;
   }
-  const headers = githubRequestHeaders(conf);
-  const response = await fetch(new Request(githubAPI, { headers }));
+  const response = await fetch(githubAPI, { headers });
   if (!response.ok) {
     const msg =
       "Error fetching repository information from GitHub. " +
@@ -69,14 +72,6 @@ export async function run(conf) {
   const editors = conf.editors.map(nameProp);
   const commenterUrls = ghCommenters ? findUserURLs(issues, comments) : [];
   const contributorUrls = ghContributors ? contributors.map(urlProp) : [];
-  const remainingRequests = await getRateLimit(conf);
-  if (commenterUrls.length + contributorUrls.length > remainingRequests) {
-    const msg =
-      `Your GitHub Repository contains ${commenterUrls.length +
-        contributorUrls.length} contributors and commenters` +
-      `but your current GitHub quota only allows ${remainingRequests} more requests. Some contributors will not show up`;
-    pub("warning", msg);
-  }
   try {
     await Promise.all(
       toHTML(commenterUrls, editors, ghCommenters, headers),
